TST: Restructure test directories
