+++ tests/base.py
@@ -5,26 +5,10 @@ import coolname
 import yaml
 
 from octue.cloud import storage
-from octue.cloud.emulators.cloud_storage import GoogleCloudStorageEmulatorTestResultModifier
 from octue.cloud.storage import GoogleCloudStorageClient
 from octue.resources import Datafile, Dataset, Manifest
-from octue.twined.cloud.emulators._pub_sub import MockTopic
-from octue.twined.cloud.emulators.service import ServicePatcher
-from octue.twined.cloud.events import OCTUE_SERVICES_TOPIC_NAME
-from tests import TEST_BUCKET_NAME, TEST_PROJECT_ID
-
-
-class TestResultModifier(GoogleCloudStorageEmulatorTestResultModifier):
-    """A test result modifier based on `GoogleCloudStorageEmulatorTestResultModifier` that also creates a mock
-    `octue.services` topic.
-    """
-
-    def startTestRun(self):
-        super().startTestRun()
-
-        with ServicePatcher():
-            self.services_topic = MockTopic(name=OCTUE_SERVICES_TOPIC_NAME, project_id=TEST_PROJECT_ID)
-            self.services_topic.create(allow_existing=True)
+from tests import TEST_BUCKET_NAME
+from tests.twined.base import TestResultModifier
 
 
 class BaseTestCase(unittest.TestCase):
