+++ tests/twined/base.py
@@ -2,6 +2,12 @@ import json
 import os
 import unittest
 
+from octue.cloud.emulators.cloud_storage import GoogleCloudStorageEmulatorTestResultModifier
+from octue.twined.cloud.emulators._pub_sub import MockTopic
+from octue.twined.cloud.emulators.service import ServicePatcher
+from octue.twined.cloud.events import OCTUE_SERVICES_TOPIC_NAME
+from tests import TEST_PROJECT_ID
+
 VALID_SCHEMA_TWINE = """
     {
         "configuration_values_schema": {
@@ -68,3 +74,16 @@ class BaseTestCase(unittest.TestCase):
         with open(valid_schema_twine_file_path, "w") as f:
             json.dump(json.loads(json_string), f)
             return valid_schema_twine_file_path
+
+
+class TestResultModifier(GoogleCloudStorageEmulatorTestResultModifier):
+    """A test result modifier based on `GoogleCloudStorageEmulatorTestResultModifier` that also creates a mock
+    `octue.services` topic.
+    """
+
+    def startTestRun(self):
+        super().startTestRun()
+
+        with ServicePatcher():
+            self.services_topic = MockTopic(name=OCTUE_SERVICES_TOPIC_NAME, project_id=TEST_PROJECT_ID)
+            self.services_topic.create(allow_existing=True)
