+++ tests/twined/cloud/emulators/test_child_emulator.py
