+++ tests/twined/cloud/events/test_attributes.py
