+++ tests/twined/cloud/pub_sub/test_logging.py
