+++ tests/twined/cloud/pub_sub/test_service.py
