+++ tests/twined/cloud/test_registry.py
