+++ tests/twined/test_app_modules/app_module_with_output_manifest/app.py
