+++ tests/twined/test_app_modules/app_using_submodule/app.py
@@ -1,7 +1,6 @@
 import logging
 
-from tests.test_app_modules.app_using_submodule.submodule import do_something
-
+from tests.twined.test_app_modules.app_using_submodule.submodule import do_something
 
 logger = logging.getLogger(__name__)
