+++ tests/twined/test_app_modules/app_using_submodule/submodule.py
@@ -1,6 +1,5 @@
 import logging
 
-
 logger = logging.getLogger(__name__)
